pyproject.toml
[project]
name = "16-livekit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "livekit-agents",
    "livekit-plugins-google",
    "livekit-plugins-cartesia",
    "livekit-plugins-deepgram",
    "livekit-plugins-noise-cancellation",
    "livekit-plugins-silero",
    "livekit-plugins-turn-detector"
]
